refactor(squad): split PT, VR, OSINT __init__.py into per-responsibility sub-modules
Review feedback on #155: the PT module reached 2034 lines after #150's
recon/save wrappers landed - structural pain that the "final-pass
sweep" framing was leaving unresolved. The next two largest agent
modules (VR at 779 lines after the args_schema sweep, OSINT at 671)
were similarly out-growing the per-module budget.

Each agent's ``__init__.py`` is now an orchestrator: import each
wrapper from its sibling sub-module, assemble ``MEMBER.tools``, and
re-export wrappers + their args_schema classes so existing consumer
imports (tests, ``crew.py``, ``tasks.py``) keep using
``from squad.<agent> import ...`` directly.

## Penetration Tester (was 2034 lines, now split across 6 files)

- ``_decorator.py`` - ``pentest_tool`` decorator, ``_PentestTool``
  Protocol, ``_parse_endpoints`` / ``_recon_from_path`` both-shapes
  adapters.
- ``probes.py`` - the 25 ``@pentest_tool`` probe wrappers (SQLi,
  SSRF, XSS, XXE, IDOR, JWT, etc.). 1188 lines; carries a
  ``# pylint: disable=C0302`` with the canonical-registry rationale.
- ``cloud.py`` - the 18 ``@cyber_tool`` cloud / infra wrappers (S3,
  Elasticsearch, Mongo, admin panels, etc.). 528 lines; same
  suppression rationale.
- ``recon.py`` - the three typed slicers over ``recon.json`` (182
  lines).
- ``findings.py`` - the ``Save Findings`` workspace writer (66
  lines).
- ``__init__.py`` - 306 lines of import + assembly + re-export.

## Vulnerability Researcher (was 779 lines, now split across 4 files)

- ``_decorator.py`` - ``research_brief_tool`` specialist wrapper and
  ``_ResearchBriefTool`` Protocol (57 lines).
- ``research.py`` - the three research-task tools (NVD CVE Lookup,
  List Programme Reports, Finalise Research) (200 lines).
- ``triage.py`` - the eight triage-task tools (List Raw Findings,
  Read Raw Finding, Calculate CVSS Score, Lookup CWE, Lookup OWASP
  Guidance, Assess Raw Finding, Discard Finding, Finalise Triage)
  plus the ``_parse_severity`` / ``_load_programme`` helpers those
  tools share (551 lines; canonical-registry C0302 suppression
  because ``Assess Raw Finding``'s authored schema dominates the
  line count rather than redundant code).
- ``__init__.py`` - 119 lines of orchestration.

## OSINT Analyst (was 671 lines, now split across 4 files)

- ``_helpers.py`` - the ``_load_programme`` loader used by both
  ``Run Initial Sweep`` (inline in discovery) and ``Finalise Recon``
  / hostname probes that need a parsed Programme.
- ``discovery.py`` - sweep + slicers + passive expansion (cert
  transparency, historical URLs, LLM endpoint detection) + active
  hostname probes + takeover detection (394 lines).
- ``curation.py`` - lookup (CWE / OWASP) + ``Annotate Host`` +
  ``Uncovered Hosts`` + ``Finalise Recon`` (255 lines).
- ``__init__.py`` - 119 lines of orchestration.

## Test patch updates

``mock.patch`` targets the name where it is LOOKED UP, not where it
is defined. The split moved imports out of each agent's
``__init__.py`` and into the new sub-modules, so existing test
patches like ``patch("squad.penetration_tester.check_ssrf", ...)``
no longer hit the wrapper's local name. Updated all three test files
(``test_squad_penetration_tester.py``, ``test_squad_vulnerability_-
researcher.py``, ``test_squad_osint_analyst.py``) to reference the
canonical submodule path - e.g. ``squad.penetration_tester.probes.-
check_ssrf``, ``squad.vulnerability_researcher.research.lookup_cve``,
``squad.osint_analyst.discovery.run_recon``. The OSINT
``_patch_programme`` test helper now targets
``squad.osint_analyst._helpers`` because that is where
``_load_programme`` (and its ``http`` / ``h1`` references) live now.

Full CI parity: ruff clean, ruff format clean, mypy clean (161 source
files - up from 150 before the split), pytest 1508 unit tests pass at
91.90% combined line + branch coverage. Pylint 9.94/10 (only three
pre-existing R0913/R0914/R0917 findings on functions that already
exceeded the cap before #150 touched them; ``C0302`` is now sweep-
wide clean for every file that does not carry an explicit
suppression with rationale).

No behaviour changes - this is a pure file-organisation refactor and
every public API (``MEMBER``, every wrapper, every args_schema class)
remains importable from the same ``squad.<agent>`` path.
