+++ squad/penetration_tester/_decorator.py
@@ -1,0 +1,90 @@
+"""
+squad/penetration_tester/_decorator.py - Pentest-specialised tool decorator
+and the both-shapes adapters every PT wrapper composes with.
+
+Lives in its own submodule (extracted from the agent's ``__init__.py``
+in the PT sub-package split) so the decorator + protocol + helpers can
+be referenced from ``probes.py`` / ``cloud.py`` / ``recon.py`` /
+``findings.py`` without dragging the registry's import surface into
+every wrapper file.
+"""
+
+from collections.abc import Callable
+from typing import Protocol, cast
+
+from pydantic import BaseModel
+
+from models import Endpoint, RawFinding, ReconResult
+from squad import CrewAITool, cyber_tool
+from tools import http
+
+_PentestFn = Callable[..., list[RawFinding]]
+
+
+class _PentestTool(CrewAITool, Protocol):
+    """Runtime shape of a CrewAI @tool: callable that also exposes .func.
+
+    Inherits ``name`` / ``description`` / ``func`` from ``CrewAITool`` and
+    adds the call signature so callers (and the pentest factory) keep the
+    ``list[RawFinding]`` return type that the squad contract test enforces.
+    """
+
+    def __call__(self, *args: object, **kwargs: object) -> list[RawFinding]: ...
+
+
+def pentest_tool(
+    name: str,
+    *,
+    check_fn: object = None,
+    args_schema: type[BaseModel],
+) -> Callable[[_PentestFn], _PentestTool]:
+    """Pentest-specialised wrapper. Composes ``cyber_tool`` plus OWASP injection.
+
+    ``cyber_tool`` handles the schema override; ``pentest_tool`` layers the
+    OWASP categories from ``check_fn.owasp_categories`` into the agent-facing
+    docstring. Every pentest probe goes through this wrapper, never bare
+    ``@tool`` and never raw ``@cyber_tool`` - the OWASP framing is what the
+    PT agent's role.md teaches it to reason against.
+    """
+
+    def decorator(fn: _PentestFn) -> _PentestTool:
+        if check_fn is not None:
+            cats = getattr(check_fn, "owasp_categories", ())
+            if cats:
+                lines = "\n".join(f"  - {c}" for c in cats)
+                fn.__doc__ = (fn.__doc__ or "").rstrip() + f"\n\nOWASP Top 10:\n{lines}\n"
+        wrapped = cyber_tool(name, args_schema=args_schema)(fn)
+        return cast(_PentestTool, wrapped)
+
+    return decorator
+
+
+def _parse_endpoints(endpoints: list[Endpoint]) -> list[Endpoint]:
+    """Re-validate the endpoint list before handing it to a probe.
+
+    The wrapper signature is ``list[Endpoint]`` so the agent-facing
+    ``args_schema`` shows a typed schema, but CrewAI's
+    ``args_schema.model_validate(...).model_dump()`` pass leaves the
+    runtime value as ``list[dict]`` by the time the wrapper body runs.
+    ``model_validate`` accepts both shapes - it returns the same Endpoint
+    when given an instance and constructs one when given a dict - so this
+    single helper is the both-shapes adapter for every probe wrapper.
+    """
+    return [Endpoint.model_validate(e) for e in endpoints]
+
+
+def _recon_from_path(recon_path: str) -> ReconResult:
+    """Read a serialised ReconResult from disk and seed the http programme context.
+
+    Centralised here so any wrapper that loads a ReconResult propagates the
+    programme handle into outbound User-Agent headers without each call site
+    having to remember the http.set_programme(...) call. ``recon_path`` is a
+    relative path under the run directory.
+    """
+    from tools.workspace import resolve_run_path
+
+    recon = ReconResult.model_validate_json(
+        resolve_run_path(recon_path).read_text(encoding="utf-8")
+    )
+    http.set_programme(recon.programme.handle)
+    return recon
