+++ squad/vulnerability_researcher/__init__.py
@@ -1,762 +1,65 @@
-"""Vulnerability Researcher - CVE intel, attack planning, and findings triage."""
+"""Vulnerability Researcher - CVE intel, attack planning, and findings triage.
 
-from collections.abc import Callable
-from datetime import UTC, datetime
+The agent's tools split across sibling modules so each file owns one
+cohesive responsibility:
+
+- ``_decorator`` - the ``research_brief_tool`` specialist wrapper and
+  its ``_ResearchBriefTool`` Protocol.
+- ``research`` - the three research-task tools the VR runs before the
+  Penetration Tester (NVD CVE Lookup, List Programme Reports, Finalise
+  Research).
+- ``triage`` - the eight triage-task tools the VR runs after the PT
+  (List Raw Findings, Read Raw Finding, Calculate CVSS Score, Lookup
+  CWE, Lookup OWASP Guidance, Assess Raw Finding, Discard Finding,
+  Finalise Triage) plus the ``_parse_severity`` / ``_load_programme``
+  helpers those tools share.
+
+This module imports each wrapper, assembles ``MEMBER.tools``, and re-
+exports both the wrappers and their args_schema classes so existing
+consumers (tests, ``crew.py``, the contract tests in
+``tests/squad/vulnerability_researcher/test_args_schemas.py``) keep
+importing from ``squad.vulnerability_researcher`` directly.
+"""
+
 from pathlib import Path
-from typing import Protocol, cast
 
-from pydantic import BaseModel, Field
-
-from models import (
-    CveEntry,
-    ProgrammeReportSummary,
-    RawFinding,
-    RawFindingSummary,
-    Severity,
-)
-from models.attack import (
-    AttackPlan,
-    AttackPlanFinalisationError,
-    AttackPlanItem,
-)
-from models.h1 import Programme
 from squad import (
-    CrewAITool,
     SquadMember,
-    cyber_tool,
     read_attack_plan_tool,
     read_run_file_tool,
     read_run_filelist_tool,
 )
-from tools import http
-from tools.cwe_data import CWEEntry
-from tools.cwe_data import lookup as cwe_lookup
-from tools.h1_api import h1
-from tools.owasp_data import OWASPEntry
-from tools.owasp_data import lookup as owasp_lookup
-from tools.pentest import lookup_cve
-from tools.report_tools import calculate_cvss_score
-from tools.research_tools import finalise_research
-from tools.research_vocab import compose_research_brief_doc
-from tools.triage_tools import (
-    AssessmentResult,
-    DiscardEntry,
-    DiscardReason,
-    DiscardResult,
-    SeverityDecision,
-    TriageAssessment,
-    TriageFinalisationError,
-    load_raw_findings,
-    save_assessment,
-    save_discard,
+from squad.vulnerability_researcher._decorator import (
+    _ResearchBriefFn,
+    _ResearchBriefTool,
+    research_brief_tool,
 )
-from tools.triage_tools import (
-    finalise_triage as finalise_triage_impl,
+from squad.vulnerability_researcher.research import (
+    _FinaliseResearchArgs,
+    _NvdCveLookupArgs,
+    _VrListProgrammeReportsArgs,
+    finalise_research_tool,
+    list_programme_reports_tool,
+    lookup_cve_tool,
 )
-from tools.triage_tools import (
-    validate_assessment as validate_assessment_impl,
+from squad.vulnerability_researcher.triage import (
+    _AssessRawFindingArgs,
+    _DiscardFindingArgs,
+    _FinaliseTriageArgs,
+    _ListRawFindingsArgs,
+    _ReadRawFindingArgs,
+    _VrCalculateCvssArgs,
+    _VrLookupCweArgs,
+    _VrLookupOwaspArgs,
+    assess_finding_tool,
+    calculate_cvss_tool,
+    discard_finding_tool,
+    finalise_triage_tool,
+    list_raw_findings_tool,
+    lookup_cwe_tool,
+    lookup_owasp_tool,
+    read_raw_finding_tool,
 )
-from tools.workspace import resolve_run_path
-
-_ResearchBriefFn = Callable[..., str]
-
-
-class _ResearchBriefTool(CrewAITool, Protocol):
-    """Runtime shape of a CrewAI @tool: callable that exposes .func and .description.
-
-    Inherits ``name`` / ``description`` / ``func`` from ``CrewAITool`` and
-    pins the call signature to the workspace-handle ``str`` return - the
-    research brief tool writes ``attack_plan.json`` and hands the next
-    agent its filename.
-    """
-
-    def __call__(self, *args: object, **kwargs: object) -> str: ...
-
-
-def research_brief_tool(
-    name: str, *, args_schema: type[BaseModel]
-) -> Callable[[_ResearchBriefFn], _ResearchBriefTool]:
-    """Research-brief-specialised wrapper. Composes ``cyber_tool`` plus
-    research-vocab injection.
-
-    Mirrors the ``pentest_tool`` factory: ``cyber_tool`` handles the
-    explicit ``args_schema`` override and ``research_brief_tool`` layers
-    the ``PROBE_VOCABULARY`` / ``RECON_EVIDENCE_KINDS`` registries from
-    ``tools.research_vocab`` into the agent-facing docstring. When a new
-    exploit family (#88) or recon-evidence kind (#45) lands, only the
-    registry is touched - the tool body and prose stay untouched.
-
-    ``args_schema=`` is keyword-required (per #150) so a brief tool
-    without an explicit schema fails import; the inferred path is no
-    longer reachable.
-    """
-
-    def decorator(fn: _ResearchBriefFn) -> _ResearchBriefTool:
-        fn.__doc__ = compose_research_brief_doc(fn.__doc__ or "")
-        wrapped = cyber_tool(name, args_schema=args_schema)(fn)
-        return cast(_ResearchBriefTool, wrapped)
-
-    return decorator
-
-
-# Research-task tools (run before the Penetration Tester)
-
-
-class _NvdCveLookupArgs(BaseModel):
-    """Explicit args_schema for the NVD CVE Lookup tool (#150)."""
-
-    keyword: str = Field(
-        description=(
-            "Free-text query against the NVD CVE index. Pass a"
-            " vulnerability class (e.g. ``SSRF``, ``deserialisation``) or"
-            " a detected technology / version pair (e.g. ``Spring Boot"
-            " 2.3``, ``Apache Struts 2.5.16``) drawn from the OSINT"
-            " Analyst's recon. Specific is better than broad: ``spring"
-            " 2.3`` returns the published Spring CVEs (incl."
-            " Spring4Shell), where ``spring`` floods the result with"
-            " unrelated matches and the PT chases noise."
-        ),
-    )
-
-
-@cyber_tool("NVD CVE Lookup", args_schema=_NvdCveLookupArgs)
-def lookup_cve_tool(keyword: str) -> list[CveEntry]:
-    """
-    Search the NVD for CVEs matching a vulnerability class or technology
-    keyword. Use this to identify known-vulnerable versions of the
-    technologies the OSINT Analyst detected, so the Penetration Tester
-    can run targeted probes instead of shotgun scans.
-    """
-    return [CveEntry(**r) for r in lookup_cve(keyword)]
-
-
-class _VrListProgrammeReportsArgs(BaseModel):
-    """Explicit args_schema for the VR's List Programme Reports tool (#150)."""
-
-    programme_handle: str = Field(
-        description=(
-            "Exact HackerOne programme handle (lowercase, no slashes, no"
-            " spaces) the report list is filtered by - the H1 API uses"
-            " the handle as the authoritative key. A mis-cased or wrong"
-            " handle returns 404 / an empty list and the VR plans against"
-            " an empty prior-art surface instead of the real one."
-        ),
-    )
-    page_size: int = Field(
-        default=25,
-        description=(
-            "Number of reports to pull per H1 API page (default 25)."
-            " Used to right-size the prior-art sample before drafting"
-            " the attack plan - 25 is enough to spot the dominant vuln"
-            " classes without paying for a full sweep."
-        ),
-    )
-
-
-@cyber_tool("List Programme Reports", args_schema=_VrListProgrammeReportsArgs)
-def list_programme_reports_tool(
-    programme_handle: str, page_size: int = 25
-) -> list[ProgrammeReportSummary]:
-    """
-    List recent public reports filed against this programme. Use this to
-    understand what vulnerability classes others have already found: spot
-    Mediums that could be escalated to Critical with a different payload or
-    attack chain, identify gaps in coverage (e.g. no SSRF reports despite a
-    rich internal surface), and flag areas defenders have already hardened.
-    Returns report id, title, severity, and state.
-    """
-    http.set_programme(programme_handle)
-    reports = h1.list_reports(programme_handle, page_size=page_size)
-    return [
-        ProgrammeReportSummary(
-            report_id=r.get("id"),
-            title=r.get("attributes", {}).get("title"),
-            severity=r.get("attributes", {}).get("severity_rating"),
-            state=r.get("attributes", {}).get("state"),
-        )
-        for r in reports
-    ]
-
-
-class _FinaliseResearchArgs(BaseModel):
-    """Explicit args_schema for the VR's Finalise Research tool (#150).
-
-    The attack plan is the contract the Penetration Tester reads to
-    decide what probes to run; a mis-shaped item either refuses (via
-    ``AttackPlanItem``'s own Pydantic validation, upstream of the
-    wrapper body) or via ``validate_attack_plan``'s hard checks. Typing
-    ``items`` as ``list[AttackPlanItem]`` (rather than ``list[dict]``)
-    makes the contract the agent reads identical to the contract the
-    persisted artefact obeys.
-    """
-
-    programme_handle: str = Field(
-        description=(
-            "Exact HackerOne programme handle the plan is drafted"
-            " against - stamped onto the AttackPlan and re-read by the"
-            " PT's ``Read Attack Plan``. Must match the handle the OSINT"
-            " Analyst recon was run against; a mismatch corrupts the"
-            " handoff."
-        ),
-    )
-    items: list[AttackPlanItem] = Field(
-        description=(
-            "Typed ``AttackPlanItem`` list, one entry per probe-target"
-            " hypothesis. Each item carries ``probe`` (canonical probe"
-            " name from the Probe vocabulary appended to this tool's"
-            " docstring), ``target`` (hostname or URL drawn from"
-            " recon), ``expected_ceiling`` (a ``Severity`` enum:"
-            " ``critical`` / ``high`` / ``medium`` / ``low`` /"
-            " ``informational`` - unknown values reject upstream),"
-            " ``rationale`` (1-2 sentences naming the why and what to"
-            " look for), and ``recon_evidence`` (non-empty list of"
-            " recon signal references drawn from the Recon evidence"
-            " kinds vocabulary; whitespace-only entries are stripped"
-            " on construction so a list of empties trips the empty-"
-            "list refusal in ``validate_attack_plan``). CrewAI hands"
-            " the wrapper ``list[dict]`` from the LLM JSON; the body"
-            " re-validates each entry as ``AttackPlanItem`` so the"
-            " persisted artefact is the canonical typed shape the PT's"
-            " ``Read Attack Plan`` consumes."
-        ),
-    )
-
-
-@research_brief_tool("Finalise Research", args_schema=_FinaliseResearchArgs)
-def finalise_research_tool(
-    programme_handle: str,
-    items: list[AttackPlanItem],
-) -> str:
-    """
-    Persist the research-task attack plan to ``attack_plan.json`` for the
-    Penetration Tester (who reads it via Read Attack Plan) and for triage
-    re-load. ``items`` is a typed ``list[AttackPlanItem]``, one entry per
-    probe-target hypothesis:
-
-      - probe: a canonical probe name. See "Probe vocabulary" below.
-      - target: hostname or URL drawn from recon
-      - expected_ceiling: a Severity enum (critical / high / medium /
-        low / informational) - the ceiling the probe could reach
-      - rationale: 1-2 sentence why and what to look for
-      - recon_evidence: list of recon signal references that justified
-        the probe. See "Recon evidence kinds" below. Must not be empty
-        (whitespace-only entries are stripped on construction).
-
-    Refuses if the plan has no items, if any item is missing one of
-    probe / target / rationale, or if any item's recon_evidence list
-    is empty after stripping. Returns the bare filename
-    ``attack_plan.json`` on success.
-    """
-    if not programme_handle:
-        raise ValueError("programme_handle is required")
-    # CrewAI args-schema validation produces list[dict] from the LLM JSON
-    # before invoking us; re-validate via the both-shapes adapter so the
-    # body accepts either a model instance (test invocations) or a dict
-    # (the LLM path) and the persisted artefact is the canonical typed
-    # shape downstream consumers depend on.
-    plan_items = [AttackPlanItem.model_validate(item) for item in items]
-    plan = AttackPlan(
-        programme_handle=programme_handle,
-        drafted_at=datetime.now(UTC),
-        items=plan_items,
-    )
-    try:
-        path = finalise_research(plan)
-    except AttackPlanFinalisationError as exc:
-        raise ValueError(str(exc)) from exc
-    return path.name
-
-
-# Triage-task tools (run after the Penetration Tester)
-
-
-class _ListRawFindingsArgs(BaseModel):
-    """Explicit args_schema for the VR's List Raw Findings tool (#150)."""
-
-    findings_path: str = Field(
-        default="findings.json",
-        description=(
-            "Relative path to the PT's raw findings artefact. Defaults"
-            " to ``findings.json`` (the canonical name written by ``Save"
-            " Findings``). Override only when re-inspecting findings"
-            " produced by a non-default writer; the typical triage call"
-            " accepts the default."
-        ),
-    )
-
-
-@cyber_tool("List Raw Findings", args_schema=_ListRawFindingsArgs)
-def list_raw_findings_tool(findings_path: str = "findings.json") -> list[RawFindingSummary]:
-    """
-    Compact summary of every raw finding in findings.json: returns
-    {index, title, vuln_class, target, severity_hint, evidence_bytes} per
-    finding, without dumping the verbose evidence payload. Use this as the
-    inventory step before deciding which findings to Assess and which to
-    Discard.
-    """
-    raw = load_raw_findings(resolve_run_path(findings_path))
-    return [
-        RawFindingSummary(
-            index=i,
-            title=f.title,
-            vuln_class=f.vuln_class,
-            target=f.target,
-            severity_hint=f.severity_hint,
-            evidence_bytes=len(f.evidence),
-        )
-        for i, f in enumerate(raw)
-    ]
-
-
-class _ReadRawFindingArgs(BaseModel):
-    """Explicit args_schema for the VR's Read Raw Finding tool (#150)."""
-
-    index: int = Field(
-        description=(
-            "Zero-based index into the findings.json array; pair with the"
-            " indices ``List Raw Findings`` returns. Out-of-range values"
-            " (negative, >= len(findings)) refuse explicitly - the"
-            " wrapper does not silently clip."
-        ),
-    )
-    findings_path: str = Field(
-        default="findings.json",
-        description=(
-            "Relative path to the PT's raw findings artefact (defaults to"
-            " ``findings.json``). Same semantics as on ``List Raw"
-            " Findings``."
-        ),
-    )
-
-
-@cyber_tool("Read Raw Finding", args_schema=_ReadRawFindingArgs)
-def read_raw_finding_tool(index: int, findings_path: str = "findings.json") -> RawFinding:
-    """
-    Return the full raw finding at ``index`` from findings.json including its
-    evidence payload (which can be large). Pair with List Raw Findings to
-    pull one finding at a time rather than loading the whole file.
-    """
-    raw = load_raw_findings(resolve_run_path(findings_path))
-    if index < 0 or index >= len(raw):
-        raise ValueError(f"index {index} out of range (findings.json has {len(raw)} entries)")
-    return raw[index]
-
-
-class _VrCalculateCvssArgs(BaseModel):
-    """Explicit args_schema for the VR's Calculate CVSS Score tool (#150).
-
-    The CVSS vector format is documented in the field description so
-    the LLM picks the right shape upstream of the call. Triage's
-    quality gate verifies that the recomputed score matches the
-    declared severity band; a wrong vector here flips the verdict on
-    the finding's severity (an inflated AV:N where the impact is
-    network-local turns a Low into a High and lands the wrong
-    bounty band downstream).
-    """
-
-    vector: str = Field(
-        description=(
-            "Full CVSS 3.1 vector string in the canonical format"
-            " ``CVSS:3.1/AV:<n>/AC:<n>/PR:<n>/UI:<n>/S:<n>/C:<n>/I:<n>"
-            "/A:<n>`` (e.g. ``CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H"
-            "/A:H``). Each metric must use the documented short codes;"
-            " a malformed vector raises rather than silently scoring 0."
-            " Use this instead of guessing the score - ``Assess Raw"
-            " Finding`` re-runs the same compute and refuses if the"
-            " declared severity does not match the computed band."
-        ),
-    )
-
-
-@cyber_tool("Calculate CVSS Score", args_schema=_VrCalculateCvssArgs)
-def calculate_cvss_tool(vector: str) -> float:
-    """
-    Compute the CVSS 3.1 base score from a vector string such as
-    CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H. Returns a value in 0.0-10.0.
-    Use this instead of guessing the score; Assess Finding verifies that
-    cvss_score matches the computed value before accepting.
-    """
-    return calculate_cvss_score(vector)
-
-
-class _VrLookupCweArgs(BaseModel):
-    """Explicit args_schema for the VR's Lookup CWE tool (#150)."""
-
-    query: str = Field(
-        description=(
-            "Free-text query against the Common Weakness Enumeration"
-            " index. Matches CWE id, name, or description"
-            " (case-insensitive substring). Use to find the canonical"
-            " MITRE definition for the vuln class on a finding before"
-            " writing the remediation - chain to ``Lookup OWASP"
-            " Guidance`` via the matching cheat-sheet topic on the"
-            " returned entry."
-        ),
-    )
-
-
-@cyber_tool("Lookup CWE", args_schema=_VrLookupCweArgs)
-def lookup_cwe_tool(query: str) -> list[CWEEntry]:
-    """
-    Find Common Weakness Enumeration entries that match a query. Useful when
-    writing the remediation section to cite the canonical MITRE definition
-    and chain to the matching OWASP cheat-sheet via Lookup OWASP Guidance.
-    """
-    return list(cwe_lookup(query))
-
-
-class _VrLookupOwaspArgs(BaseModel):
-    """Explicit args_schema for the VR's Lookup OWASP Guidance tool (#150)."""
-
-    query: str = Field(
-        description=(
-            "Free-text query against the OWASP Cheat Sheet index. Matches"
-            " topic slug, title, or short alias (e.g. ``ssrf``, ``csrf``,"
-            " ``sqli``). Use to surface canonical guidance for the"
-            " remediation citation - the returned entry's URL is the"
-            " ``cheatsheetseries.owasp.org`` link the assessment is"
-            " expected to cite."
-        ),
-    )
-
-
-@cyber_tool("Lookup OWASP Guidance", args_schema=_VrLookupOwaspArgs)
-def lookup_owasp_tool(query: str) -> list[OWASPEntry]:
-    """
-    Find OWASP Cheat Sheet entries that match a query (topic slug, title, or
-    short alias such as "ssrf", "csrf", "sqli"). Returns title, paste-friendly
-    key_principles, and the canonical cheatsheetseries.owasp.org URL to cite
-    in the remediation section.
-    """
-    return list(owasp_lookup(query))
-
-
-class _AssessRawFindingArgs(BaseModel):
-    """Explicit args_schema for the VR's Assess Raw Finding tool (#150).
-
-    The longest and highest-stakes contract on the VR: every authored
-    field below feeds the assessment quality gate. The per-field
-    descriptions are how the LLM learns the gate's grammar before it
-    calls the tool, so wrong-shape assessments refuse upstream rather
-    than after the body has written a half-validated artefact.
-    """
-
-    finding_index: int = Field(
-        description=(
-            "Zero-based index into findings.json - matches the indices"
-            " ``List Raw Findings`` returns. Carries target / vuln_class"
-            " / severity_hint forward from the PT's finding so the"
-            " assessment cannot contradict the raw finding it derives"
-            " from. Out-of-range refuses explicitly."
-        ),
-    )
-    severity_decision: SeverityDecision = Field(
-        description=(
-            "How this severity call relates to the PT's"
-            " ``severity_hint``: one of ``keep`` (the VR agrees with the"
-            " PT), ``raise`` (the VR's analysis warrants a higher band -"
-            " explain why in ``severity_rationale``), or ``lower``"
-            " (the VR's analysis warrants a lower band). Typed as the"
-            " ``SeverityDecision`` enum so unknown values reject"
-            " upstream of the wrapper body."
-        ),
-    )
-    severity: Severity = Field(
-        description=(
-            "Final severity band after the decision, typed as the"
-            " ``Severity`` enum (``critical`` / ``high`` / ``medium`` /"
-            " ``low`` / ``informational``). Unknown values reject"
-            " upstream. The quality gate then verifies the band matches"
-            " the recomputed CVSS score - mis-banding refuses at the"
-            " body level."
-        ),
-    )
-    severity_rationale: str = Field(
-        description=(
-            "1-2 sentences justifying the severity call. Required even"
-            " when ``severity_decision`` is ``keep``; the rationale is"
-            " surfaced in the briefing to the Technical Author."
-        ),
-    )
-    cvss_vector: str = Field(
-        description=(
-            "Full CVSS 3.1 vector string"
-            " (``CVSS:3.1/AV:<n>/AC:<n>/PR:<n>/UI:<n>/S:<n>/C:<n>/I:<n>"
-            "/A:<n>``). ``cvss_score`` is recomputed from this vector and"
-            " verified against the declared ``severity`` band. Use"
-            " ``Calculate CVSS Score`` first if you want to double-check"
-            " the score before submitting."
-        ),
-    )
-    title: str = Field(
-        description=(
-            "Short, descriptive title. The Technical Author may rewrite"
-            " for H1 format, but keep it informative enough that the"
-            " briefing reads usefully on its own."
-        ),
-    )
-    description: str = Field(
-        description=(
-            "Root-cause explanation aimed at a developer audience. The"
-            " quality gate refuses overly short descriptions - give the"
-            " reader the WHY, not just the what."
-        ),
-    )
-    steps_to_reproduce: list[str] = Field(
-        description=(
-            "Numbered steps a triager can follow verbatim to reproduce"
-            " the issue. The Technical Author refines them into the"
-            " report; ship them concrete enough that the refinement is"
-            " editing, not authoring from scratch."
-        ),
-    )
-    impact: str = Field(
-        description=(
-            "Specific - name the data / system at risk and the worst"
-            " realistic outcome. Hand-wavy language (``could"
-            " compromise``, ``potential``) is rejected by the quality"
-            " gate; the report's impact section is what the programme"
-            " uses to band severity."
-        ),
-    )
-    remediation: str = Field(
-        description=(
-            "Actionable fix paired with an OWASP cheat-sheet or CWE URL"
-            " citation. The gate refuses remediations that name no fix"
-            " or carry no citation - use ``Lookup CWE`` / ``Lookup OWASP"
-            " Guidance`` to find the matching URL."
-        ),
-    )
-    findings_path: str = Field(
-        default="findings.json",
-        description=(
-            "Relative path to the raw findings artefact (defaults to"
-            " ``findings.json``). Same semantics as on the other"
-            " findings-reading tools."
-        ),
-    )
-    programme_handle: str | None = Field(
-        default=None,
-        description=(
-            "Exact HackerOne programme handle the scope / floor gates"
-            " check against. Required - the gate refuses out-of-scope"
-            " findings and findings below the configured min-severity"
-            " floor, and both checks need a parsed Programme."
-        ),
-    )
-
-
-@cyber_tool("Assess Raw Finding", args_schema=_AssessRawFindingArgs)
-# CrewAI builds the tool's JSON schema from this signature; every authored
-# field has to be named so the LLM can fill them. Collapsing into a single
-# dict argument would force the agent to guess valid keys.
-# pylint: disable=R0913,R0917
-def assess_finding_tool(
-    finding_index: int,
-    severity_decision: str,
-    severity: str,
-    severity_rationale: str,
-    cvss_vector: str,
-    title: str,
-    description: str,
-    steps_to_reproduce: list[str],
-    impact: str,
-    remediation: str,
-    findings_path: str = "findings.json",
-    programme_handle: str | None = None,
-) -> AssessmentResult:
-    """
-    Author one triage assessment for the raw finding at ``finding_index`` and
-    run the quality gate. The carried fields (target, vuln_class,
-    severity_hint) are pulled from findings.json so you cannot contradict the
-    Penetration Tester.
-
-    Inputs you author:
-      - severity_decision: "keep" | "raise" | "lower" relative to the PT hint
-      - severity: the final severity after your decision
-      - severity_rationale: 1-2 sentences justifying the call
-      - cvss_vector: full CVSS 3.1 vector; cvss_score is recomputed and
-        verified against your declared severity band
-      - title: TA may rewrite, but keep it descriptive
-      - description: root-cause explanation aimed at a developer audience
-      - steps_to_reproduce: numbered steps the TA can refine into the report
-      - impact: specific - name the data/system at risk and worst outcome.
-        Hand-wavy language ("could compromise", "potential") is rejected.
-      - remediation: actionable fix with an OWASP or CWE URL citation
-
-    Out-of-scope or below-floor findings are rejected: use Discard Finding
-    instead. Returns an AssessmentResult with the relative assessment path
-    and a TriageValidationReport; when validation.ok is false, re-run with
-    the issues addressed.
-    """
-    raw = load_raw_findings(resolve_run_path(findings_path))
-    if finding_index < 0 or finding_index >= len(raw):
-        raise ValueError(f"finding_index {finding_index} out of range")
-    finding = raw[finding_index]
-
-    programme = _load_programme(programme_handle)
-
-    assessment = TriageAssessment(
-        finding_index=finding_index,
-        target=finding.target,
-        vuln_class=finding.vuln_class,
-        severity_hint=finding.severity_hint,
-        severity_decision=SeverityDecision(severity_decision),
-        severity=_parse_severity(severity),
-        severity_rationale=severity_rationale.strip(),
-        cvss_vector=cvss_vector.strip(),
-        cvss_score=calculate_cvss_score(cvss_vector.strip()),
-        title=title.strip(),
-        description=description.strip(),
-        steps_to_reproduce=[s.strip() for s in steps_to_reproduce],
-        impact=impact.strip(),
-        remediation=remediation.strip(),
-    )
-    path = save_assessment(assessment)
-    return AssessmentResult(
-        path=str(path.relative_to(path.parents[1])),
-        validation=validate_assessment_impl(assessment, finding, programme),
-    )
-
-
-class _DiscardFindingArgs(BaseModel):
-    """Explicit args_schema for the VR's Discard Finding tool (#150)."""
-
-    finding_index: int = Field(
-        description=(
-            "Zero-based index of the raw finding to discard - matches the"
-            " indices ``List Raw Findings`` returns. Out-of-range refuses"
-            " explicitly."
-        ),
-    )
-    reason: str = Field(
-        description=(
-            "Canonical discard reason: one of ``out_of_scope``,"
-            " ``below_floor``, ``false_positive``, ``duplicate``,"
-            " ``non_exploitable``. The wrapper coerces to"
-            " ``DiscardReason``; unknown values reject. Each reason maps"
-            " to a different downstream behaviour (a ``duplicate`` discard"
-            " surfaces differently in the briefing than a"
-            " ``false_positive``), so pick the most specific match."
-        ),
-    )
-    rationale: str = Field(
-        description=(
-            "One-sentence justification recorded on the discard entry and"
-            " surfaced in the briefing to the Technical Author. Must be"
-            " at least 10 characters - the wrapper refuses thinner"
-            " rationales because the next agent needs the why."
-        ),
-    )
-    findings_path: str = Field(
-        default="findings.json",
-        description=(
-            "Relative path to the raw findings artefact (defaults to"
-            " ``findings.json``). Same semantics as on the other"
-            " findings-reading tools."
-        ),
-    )
-
-
-@cyber_tool("Discard Finding", args_schema=_DiscardFindingArgs)
-def discard_finding_tool(
-    finding_index: int,
-    reason: str,
-    rationale: str,
-    findings_path: str = "findings.json",
-) -> DiscardResult:
-    """
-    Mark a raw finding as not eligible for verified.json. ``reason`` must be
-    one of: out_of_scope, below_floor, false_positive, duplicate,
-    non_exploitable. ``rationale`` is a one-sentence justification recorded
-    on the discard entry and surfaced in the briefing to the Technical
-    Author. Returns a DiscardResult with the relative discard path.
-    """
-    raw = load_raw_findings(resolve_run_path(findings_path))
-    if finding_index < 0 or finding_index >= len(raw):
-        raise ValueError(f"finding_index {finding_index} out of range")
-    finding = raw[finding_index]
-    if len(rationale.strip()) < 10:
-        raise ValueError("rationale must be at least 10 chars - explain why this is discarded")
-    entry = DiscardEntry(
-        finding_index=finding_index,
-        target=finding.target,
-        vuln_class=finding.vuln_class,
-        severity_hint=finding.severity_hint,
-        reason=DiscardReason(reason),
-        rationale=rationale.strip(),
-    )
-    path = save_discard(entry)
-    return DiscardResult(path=str(path.relative_to(path.parents[1])))
-
-
-class _FinaliseTriageArgs(BaseModel):
-    """Explicit args_schema for the VR's Finalise Triage tool (#150)."""
-
-    programme_handle: str = Field(
-        description=(
-            "Exact HackerOne programme handle the scope / floor gates"
-            " re-check against before consolidation. Must match the"
-            " handle the PT's findings were produced under; a mismatch"
-            " would re-check scope against the wrong programme and let"
-            " out-of-scope findings slip through."
-        ),
-    )
-    findings_path: str = Field(
-        default="findings.json",
-        description=(
-            "Relative path to the raw findings artefact (defaults to"
-            " ``findings.json``). Same semantics as on the other"
-            " findings-reading tools."
-        ),
-    )
-
-
-@cyber_tool("Finalise Triage", args_schema=_FinaliseTriageArgs)
-def finalise_triage_tool(
-    programme_handle: str,
-    findings_path: str = "findings.json",
-) -> str:
-    """
-    Consolidate every assessment into verified.json for the Technical Author.
-    Refuses if any raw finding has neither an assessment nor a discard, or
-    if any assessment has unresolved validation errors. The error message
-    lists exactly which finding / section needs work.
-
-    Returns the bare filename "verified.json" on success.
-    """
-    raw = load_raw_findings(resolve_run_path(findings_path))
-    programme = _load_programme(programme_handle)
-    try:
-        path = finalise_triage_impl(programme, raw)
-    except TriageFinalisationError as exc:
-        raise ValueError(str(exc)) from exc
-    return path.name
-
-
-# Helpers
-
-
-def _parse_severity(severity: str) -> Severity:
-    try:
-        return Severity(severity.lower())
-    except ValueError as exc:
-        raise ValueError(
-            f"severity must be one of {[s.value for s in Severity]}, got {severity!r}"
-        ) from exc
-
-
-def _load_programme(programme_handle: str | None) -> Programme:
-    """Fetch and parse the Programme. The scope/floor gates need a real
-    Programme to check against, so the triage tools call this on every
-    invocation."""
-    if not programme_handle:
-        raise ValueError("programme_handle is required")
-    http.set_programme(programme_handle)
-    policy = h1.get_programme_policy(programme_handle)
-    scope = h1.get_structured_scope(programme_handle)
-    return h1.parse_programme(policy["data"], scope)
-
 
 MEMBER = SquadMember(
     dir=Path(__file__).parent,
@@ -780,3 +83,37 @@ MEMBER = SquadMember(
         read_run_file_tool,
     ],
 )
+
+__all__ = [
+    # Public API
+    "MEMBER",
+    # Decorator surface
+    "_ResearchBriefFn",
+    "_ResearchBriefTool",
+    "research_brief_tool",
+    # Research-task wrappers
+    "finalise_research_tool",
+    "list_programme_reports_tool",
+    "lookup_cve_tool",
+    # Triage-task wrappers
+    "assess_finding_tool",
+    "calculate_cvss_tool",
+    "discard_finding_tool",
+    "finalise_triage_tool",
+    "list_raw_findings_tool",
+    "lookup_cwe_tool",
+    "lookup_owasp_tool",
+    "read_raw_finding_tool",
+    # args_schema classes (re-exported so test imports stay stable)
+    "_AssessRawFindingArgs",
+    "_DiscardFindingArgs",
+    "_FinaliseResearchArgs",
+    "_FinaliseTriageArgs",
+    "_ListRawFindingsArgs",
+    "_NvdCveLookupArgs",
+    "_ReadRawFindingArgs",
+    "_VrCalculateCvssArgs",
+    "_VrListProgrammeReportsArgs",
+    "_VrLookupCweArgs",
+    "_VrLookupOwaspArgs",
+]
