+++ squad/vulnerability_researcher/_decorator.py
@@ -1,0 +1,57 @@
+"""
+squad/vulnerability_researcher/_decorator.py - the
+``research_brief_tool`` specialist wrapper and its ``_ResearchBriefTool``
+Protocol.
+
+Lives in its own submodule (extracted from the agent's ``__init__.py``
+in the VR sub-package split) so ``research.py`` can decorate the
+research-task wrapper without pulling in every other VR import.
+"""
+
+from collections.abc import Callable
+from typing import Protocol, cast
+
+from pydantic import BaseModel
+
+from squad import CrewAITool, cyber_tool
+from tools.research_vocab import compose_research_brief_doc
+
+_ResearchBriefFn = Callable[..., str]
+
+
+class _ResearchBriefTool(CrewAITool, Protocol):
+    """Runtime shape of a CrewAI @tool: callable that exposes .func and .description.
+
+    Inherits ``name`` / ``description`` / ``func`` from ``CrewAITool`` and
+    pins the call signature to the workspace-handle ``str`` return - the
+    research brief tool writes ``attack_plan.json`` and hands the next
+    agent its filename.
+    """
+
+    def __call__(self, *args: object, **kwargs: object) -> str: ...
+
+
+def research_brief_tool(
+    name: str, *, args_schema: type[BaseModel]
+) -> Callable[[_ResearchBriefFn], _ResearchBriefTool]:
+    """Research-brief-specialised wrapper. Composes ``cyber_tool`` plus
+    research-vocab injection.
+
+    Mirrors the ``pentest_tool`` factory: ``cyber_tool`` handles the
+    explicit ``args_schema`` override and ``research_brief_tool`` layers
+    the ``PROBE_VOCABULARY`` / ``RECON_EVIDENCE_KINDS`` registries from
+    ``tools.research_vocab`` into the agent-facing docstring. When a new
+    exploit family (#88) or recon-evidence kind (#45) lands, only the
+    registry is touched - the tool body and prose stay untouched.
+
+    ``args_schema=`` is keyword-required (per #150) so a brief tool
+    without an explicit schema fails import; the inferred path is no
+    longer reachable.
+    """
+
+    def decorator(fn: _ResearchBriefFn) -> _ResearchBriefTool:
+        fn.__doc__ = compose_research_brief_doc(fn.__doc__ or "")
+        wrapped = cyber_tool(name, args_schema=args_schema)(fn)
+        return cast(_ResearchBriefTool, wrapped)
+
+    return decorator
