+++ squad/vulnerability_researcher/research.py
@@ -1,0 +1,200 @@
+"""
+squad/vulnerability_researcher/research.py - the three research-task
+tools the Vulnerability Researcher runs before the Penetration Tester.
+
+The agent uses these to turn the OSINT Analyst's recon brief into a
+typed attack plan (``attack_plan.json``) that the PT then reads via
+``Read Attack Plan``. Tools:
+
+- ``NVD CVE Lookup`` - search the NVD for CVEs matching a class /
+  technology keyword.
+- ``List Programme Reports`` - prior-art sample from H1's recent public
+  reports on the programme.
+- ``Finalise Research`` - persist the typed ``AttackPlan`` to the run
+  directory.
+"""
+
+from datetime import UTC, datetime
+
+from pydantic import BaseModel, Field
+
+from models import CveEntry, ProgrammeReportSummary
+from models.attack import (
+    AttackPlan,
+    AttackPlanFinalisationError,
+    AttackPlanItem,
+)
+from squad import cyber_tool
+from squad.vulnerability_researcher._decorator import research_brief_tool
+from tools import http
+from tools.h1_api import h1
+from tools.pentest import lookup_cve
+from tools.research_tools import finalise_research
+
+
+class _NvdCveLookupArgs(BaseModel):
+    """Explicit args_schema for the NVD CVE Lookup tool (#150)."""
+
+    keyword: str = Field(
+        description=(
+            "Free-text query against the NVD CVE index. Pass a"
+            " vulnerability class (e.g. ``SSRF``, ``deserialisation``) or"
+            " a detected technology / version pair (e.g. ``Spring Boot"
+            " 2.3``, ``Apache Struts 2.5.16``) drawn from the OSINT"
+            " Analyst's recon. Specific is better than broad: ``spring"
+            " 2.3`` returns the published Spring CVEs (incl."
+            " Spring4Shell), where ``spring`` floods the result with"
+            " unrelated matches and the PT chases noise."
+        ),
+    )
+
+
+@cyber_tool("NVD CVE Lookup", args_schema=_NvdCveLookupArgs)
+def lookup_cve_tool(keyword: str) -> list[CveEntry]:
+    """
+    Search the NVD for CVEs matching a vulnerability class or technology
+    keyword. Use this to identify known-vulnerable versions of the
+    technologies the OSINT Analyst detected, so the Penetration Tester
+    can run targeted probes instead of shotgun scans.
+    """
+    return [CveEntry(**r) for r in lookup_cve(keyword)]
+
+
+class _VrListProgrammeReportsArgs(BaseModel):
+    """Explicit args_schema for the VR's List Programme Reports tool (#150)."""
+
+    programme_handle: str = Field(
+        description=(
+            "Exact HackerOne programme handle (lowercase, no slashes, no"
+            " spaces) the report list is filtered by - the H1 API uses"
+            " the handle as the authoritative key. A mis-cased or wrong"
+            " handle returns 404 / an empty list and the VR plans against"
+            " an empty prior-art surface instead of the real one."
+        ),
+    )
+    page_size: int = Field(
+        default=25,
+        description=(
+            "Number of reports to pull per H1 API page (default 25)."
+            " Used to right-size the prior-art sample before drafting"
+            " the attack plan - 25 is enough to spot the dominant vuln"
+            " classes without paying for a full sweep."
+        ),
+    )
+
+
+@cyber_tool("List Programme Reports", args_schema=_VrListProgrammeReportsArgs)
+def list_programme_reports_tool(
+    programme_handle: str, page_size: int = 25
+) -> list[ProgrammeReportSummary]:
+    """
+    List recent public reports filed against this programme. Use this to
+    understand what vulnerability classes others have already found: spot
+    Mediums that could be escalated to Critical with a different payload or
+    attack chain, identify gaps in coverage (e.g. no SSRF reports despite a
+    rich internal surface), and flag areas defenders have already hardened.
+    Returns report id, title, severity, and state.
+    """
+    http.set_programme(programme_handle)
+    reports = h1.list_reports(programme_handle, page_size=page_size)
+    return [
+        ProgrammeReportSummary(
+            report_id=r.get("id"),
+            title=r.get("attributes", {}).get("title"),
+            severity=r.get("attributes", {}).get("severity_rating"),
+            state=r.get("attributes", {}).get("state"),
+        )
+        for r in reports
+    ]
+
+
+class _FinaliseResearchArgs(BaseModel):
+    """Explicit args_schema for the VR's Finalise Research tool (#150).
+
+    The attack plan is the contract the Penetration Tester reads to
+    decide what probes to run; a mis-shaped item either refuses (via
+    ``AttackPlanItem``'s own Pydantic validation, upstream of the
+    wrapper body) or via ``validate_attack_plan``'s hard checks. Typing
+    ``items`` as ``list[AttackPlanItem]`` (rather than ``list[dict]``)
+    makes the contract the agent reads identical to the contract the
+    persisted artefact obeys.
+    """
+
+    programme_handle: str = Field(
+        description=(
+            "Exact HackerOne programme handle the plan is drafted"
+            " against - stamped onto the AttackPlan and re-read by the"
+            " PT's ``Read Attack Plan``. Must match the handle the OSINT"
+            " Analyst recon was run against; a mismatch corrupts the"
+            " handoff."
+        ),
+    )
+    items: list[AttackPlanItem] = Field(
+        description=(
+            "Typed ``AttackPlanItem`` list, one entry per probe-target"
+            " hypothesis. Each item carries ``probe`` (canonical probe"
+            " name from the Probe vocabulary appended to this tool's"
+            " docstring), ``target`` (hostname or URL drawn from"
+            " recon), ``expected_ceiling`` (a ``Severity`` enum:"
+            " ``critical`` / ``high`` / ``medium`` / ``low`` /"
+            " ``informational`` - unknown values reject upstream),"
+            " ``rationale`` (1-2 sentences naming the why and what to"
+            " look for), and ``recon_evidence`` (non-empty list of"
+            " recon signal references drawn from the Recon evidence"
+            " kinds vocabulary; whitespace-only entries are stripped"
+            " on construction so a list of empties trips the empty-"
+            "list refusal in ``validate_attack_plan``). CrewAI hands"
+            " the wrapper ``list[dict]`` from the LLM JSON; the body"
+            " re-validates each entry as ``AttackPlanItem`` so the"
+            " persisted artefact is the canonical typed shape the PT's"
+            " ``Read Attack Plan`` consumes."
+        ),
+    )
+
+
+@research_brief_tool("Finalise Research", args_schema=_FinaliseResearchArgs)
+def finalise_research_tool(
+    programme_handle: str,
+    items: list[AttackPlanItem],
+) -> str:
+    """
+    Persist the research-task attack plan to ``attack_plan.json`` for the
+    Penetration Tester (who reads it via Read Attack Plan) and for triage
+    re-load. ``items`` is a typed ``list[AttackPlanItem]``, one entry per
+    probe-target hypothesis:
+
+      - probe: a canonical probe name. See "Probe vocabulary" below.
+      - target: hostname or URL drawn from recon
+      - expected_ceiling: a Severity enum (critical / high / medium /
+        low / informational) - the ceiling the probe could reach
+      - rationale: 1-2 sentence why and what to look for
+      - recon_evidence: list of recon signal references that justified
+        the probe. See "Recon evidence kinds" below. Must not be empty
+        (whitespace-only entries are stripped on construction).
+
+    Refuses if the plan has no items, if any item is missing one of
+    probe / target / rationale, or if any item's recon_evidence list
+    is empty after stripping. Returns the bare filename
+    ``attack_plan.json`` on success.
+    """
+    if not programme_handle:
+        raise ValueError("programme_handle is required")
+    # CrewAI args-schema validation produces list[dict] from the LLM JSON
+    # before invoking us; re-validate via the both-shapes adapter so the
+    # body accepts either a model instance (test invocations) or a dict
+    # (the LLM path) and the persisted artefact is the canonical typed
+    # shape downstream consumers depend on.
+    plan_items = [AttackPlanItem.model_validate(item) for item in items]
+    plan = AttackPlan(
+        programme_handle=programme_handle,
+        drafted_at=datetime.now(UTC),
+        items=plan_items,
+    )
+    try:
+        path = finalise_research(plan)
+    except AttackPlanFinalisationError as exc:
+        raise ValueError(str(exc)) from exc
+    return path.name
+
+
+# Triage-task tools (run after the Penetration Tester)
