+++ tests/test_squad_vulnerability_researcher.py
@@ -29,7 +29,7 @@ class TestVulnerabilityResearcherTools:
             description="Test CVE",
         )
         with patch(
-            "squad.vulnerability_researcher.lookup_cve",
+            "squad.vulnerability_researcher.research.lookup_cve",
             return_value=[expected.model_dump()],
         ) as m:
             result = lookup_cve_tool.func("sql injection")
@@ -59,9 +59,9 @@ class TestVulnerabilityResearcherTools:
             },
         ]
         with (
-            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch("squad.vulnerability_researcher.research.http.set_programme"),
             patch(
-                "squad.vulnerability_researcher.h1.list_reports",
+                "squad.vulnerability_researcher.research.h1.list_reports",
                 return_value=reports,
             ),
         ):
@@ -77,9 +77,9 @@ class TestVulnerabilityResearcherTools:
 
         reports = [{"id": "3", "attributes": {"state": "closed"}}]
         with (
-            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch("squad.vulnerability_researcher.research.http.set_programme"),
             patch(
-                "squad.vulnerability_researcher.h1.list_reports",
+                "squad.vulnerability_researcher.research.h1.list_reports",
                 return_value=reports,
             ),
         ):
@@ -296,14 +296,14 @@ class TestVulnerabilityResearcherTools:
     @staticmethod
     def _patch_programme(programme):
         return [
-            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch("squad.vulnerability_researcher.triage.http.set_programme"),
             patch(
-                "squad.vulnerability_researcher.h1.get_programme_policy",
+                "squad.vulnerability_researcher.triage.h1.get_programme_policy",
                 return_value={"data": {}},
             ),
-            patch("squad.vulnerability_researcher.h1.get_structured_scope", return_value={}),
+            patch("squad.vulnerability_researcher.triage.h1.get_structured_scope", return_value={}),
             patch(
-                "squad.vulnerability_researcher.h1.parse_programme",
+                "squad.vulnerability_researcher.triage.h1.parse_programme",
                 return_value=programme,
             ),
         ]
